NNet horizontal/vertical (completed)

+++ src/main/java/cz/gattserver/nnet/net/NNet.java
@@ -145,7 +145,7 @@ public class NNet {
 	}
 
 	public double getSuccessRate() {
-		return successCount / triesCount * 100;
+		return (double) successCount / triesCount * 100;
 	}
 
 }
